fix(store): remove unused code
